fix(audit): low-severity hardening batch (legacy parents, fee overflow, counter drift, slash rent, snap votes)
- create_dependent_task: parent_task is now an UncheckedAccount with
  in-handler owner-check + zero-pad deserialize, so legacy 382B
  (pre-migration) parents load instead of bricking the instruction with
  AccountDeserialize. The status (Cancelled/Disputed) and same-creator
  checks moved into the handler unchanged.

- purchase_good / purchase_skill: protocol-fee multiply now uses u128
  intermediates (F-16 pattern). price.checked_mul(bps) overflowed u64
  above u64::MAX/10_000, DoS-ing purchases at huge prices.

- resolve_dispute / expire_dispute: defendant active_tasks decrement is
  now saturating_sub, matching every other decrement of that counter
  (process_worker_claim_pair). A legacy drifted counter can no longer
  brick the designated un-bricking exits with ArithmeticOverflow.

- apply_dispute_slash: the token-settlement branch returns the escrow
  PDA + ATA rent to the task CREATOR (who funded both at create_task),
  matching every other close path in the program — previously sent to
  the treasury. Adds a trailing OPTIONAL creator account (validated
  against task.creator, required only when the settlement branch runs).
  Non-canary instruction; wire-compatible for SOL callers (null).

- create_proposal: a caller-provided voting_period is now floored at the
  governance default in addition to the MAX cap — a proposer can no
  longer close voting before the electorate can react (a mainnet
  proposal already ran at 600s).

Tests: dependent-legacy-parent.test.mjs (3), dispute-counter-drift.test.mjs
(1), governance floor/cap test (1), purchase_good u128 unit test — all
revert-sensitive. Suite 307/307; 415 unit tests; clippy clean both
feature sets; canary 25-ix wire freeze intact; sdk drift green.

+++ programs/agenc-coordination/src/instructions/purchase_skill.rs
@@ -141,11 +141,13 @@ pub fn handler(ctx: Context<PurchaseSkill>, expected_price: u64) -> Result<()> {
             );
 
             // Calculate fee
-            protocol_fee = price
-                .checked_mul(config.protocol_fee_bps as u64)
+            // u128 intermediates (audit F-16): price.checked_mul(bps) overflows
+            // u64 for prices above ~9.2e15; the post-division fee always fits.
+            protocol_fee = (price as u128)
+                .checked_mul(config.protocol_fee_bps as u128)
                 .ok_or(CoordinationError::ArithmeticOverflow)?
-                .checked_div(BASIS_POINTS_DIVISOR)
-                .ok_or(CoordinationError::ArithmeticOverflow)?;
+                .checked_div(BASIS_POINTS_DIVISOR as u128)
+                .ok_or(CoordinationError::ArithmeticOverflow)? as u64;
             let author_share = price
                 .checked_sub(protocol_fee)
                 .ok_or(CoordinationError::ArithmeticOverflow)?;
@@ -231,11 +233,13 @@ pub fn handler(ctx: Context<PurchaseSkill>, expected_price: u64) -> Result<()> {
             }
         } else {
             // SOL payment path
-            protocol_fee = price
-                .checked_mul(config.protocol_fee_bps as u64)
+            // u128 intermediates (audit F-16): price.checked_mul(bps) overflows
+            // u64 for prices above ~9.2e15; the post-division fee always fits.
+            protocol_fee = (price as u128)
+                .checked_mul(config.protocol_fee_bps as u128)
                 .ok_or(CoordinationError::ArithmeticOverflow)?
-                .checked_div(BASIS_POINTS_DIVISOR)
-                .ok_or(CoordinationError::ArithmeticOverflow)?;
+                .checked_div(BASIS_POINTS_DIVISOR as u128)
+                .ok_or(CoordinationError::ArithmeticOverflow)? as u64;
             let author_share = price
                 .checked_sub(protocol_fee)
                 .ok_or(CoordinationError::ArithmeticOverflow)?;
